pyproject.toml
[tool.poetry]
name = "streamdeck_ui"
version = "2.0.16"
description = "A service, Web Interface, and UI for interacting with your computer using a Stream Deck"
authors = ["Timothy Crosley <timothy.crosley@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
streamdeck = "^0.9.3"
pillow = "^9.4.0"
pynput = "^1.7.6"
pyside6 = "^6.4.2"
CairoSVG = "^2.5.2"
filetype = "^1.0.10"
python-xlib = "^0.33"
slackclient = "^2.9.4"

[tool.poetry.dev-dependencies]
vulture = ">=1.0"
bandit = ">=1.6"
safety = ">=1.8"
isort = ">=5.3"
flake8-bugbear = ">=19.8"
black = {version = ">=18.3-alpha.0", allow-prereleases = true}
mypy = ">=0.730.0"
ipython = ">=7.7"
pytest = ">=5.0"
pytest-cov = ">=2.7"
pytest-mock = ">=1.10"
pep8-naming = ">=0.8.2"
portray = ">=1.3.0"
cruft = ">=2.2"
jupyter = "^1.0"
ipdb = "^0.12.2"
hypothesis-auto = "^1.1"
pydantic = "^1.6"
numpy = "^1.22.1"
types-pkg-resources = "^0.1.3"

[tool.poetry.scripts]
streamdeck = "streamdeck_ui.gui:start"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "hug"

[tool.portray]
output_dir = "site"
include_reference_documentation = false
extra_dirs = ["art", "images", "media", "scripts"]

[tool.portray.mkdocs]
repo_url = "https://github.com/timothycrosley/streamdeck-ui"
edit_uri = "edit/master/"
